Implemented precision cleanup and documented the behavior. mixed_precision now means: autocast only around the model forward. It defaults to bfloat16, and now supports: mixed_precision: true mixed_precision_dtype: bfloat16  # or float16 For true float64 training, use: default_dtype: float64 mixed_precision: false allow_tf32: false Changes made: Floating dataset fields from numpy, lists, and tensors now follow torch.get_default_dtype() in [AtomicData.py](/home/angiod@usi.ch/GEqTrain/geqtrain/data/AtomicData.py). Processed dataset cache now includes active dtype in [dataset.py](/home/angiod@usi.ch/GEqTrain/geqtrain/data/dataset.py), so float64 builds a separate cache. Computation-relevant buffers/parameters in radial bases, SO3 layers, readout, scale modules, recycling, embeddings, loss coeffs, and normalization now use default dtype instead of hardcoded float32. PBC edge shifts now cast to position dtype in [AtomicDataDict.py](/home/angiod@usi.ch/GEqTrain/geqtrain/data/AtomicDataDict.py). Added precision docs at [precision.md](/home/angiod@usi.ch/GEqTrain/docs/precision.md). Added tests at [test_precision_config.py](/home/angiod@usi.ch/GEqTrain/tests/test_precision_config.py). Remaining float32 casts are intentionally only for logging/diagnostics/sampler scalar bookkeeping or external AUROC handling, not model computation.

+++ geqtrain/train/_loss.py
@@ -604,7 +604,7 @@ class FocalLossBinaryAccuracy:
 
     def __call__(self, pred: dict, ref: dict, key: str, mean: bool = True, **kwargs):
         logits = pred[key]
-        target = ref[key].float()
+        target = ref[key].to(dtype=logits.dtype)
 
         bce_loss = self.bce(logits, target)
         p_t = torch.exp(-bce_loss) # This is p if target=1, and 1-p if target=0
